audit-logs-export/requirements.txt
requests==2.28.1
google-cloud-storage==2.5.0
Flask==2.1.2